Added multiline print capabilities to the 3 PrintText functions All 3 Text functions now accept multiline text, each line separate by \n Line spacing is done according to the font height. All 3 text functions now return the next "unused" line. Some minor changes (startup screen delay, remove unused includes in eclipse) with no functional impact

+++ mchf-eclipse/drivers/ui/ui_driver.h
@@ -286,7 +286,7 @@ void	UiDriver_LcdBlankingStartTimer(void);
 void	UiDriver_ShowDebugText(const char*);
 void 	UiDriver_ChangeTuningStep(uchar is_up);
 void	UiDriver_UpdateDisplayAfterParamChange();
-void    UiDriver_ShowStartUpScreen(ulong hold_time);
+void    UiDriver_ShowStartUpScreen(uint32_t hold_time);
 void    UiDriver_DoCrossCheck(char cross[],char* xt_corr, char* yt_corr);
 void    UiDriver_ToggleVfoAB();
 void    UiDriver_SetSplitMode(bool mode_active);
